+++ R/pvlrt-function.R
@@ -1,0 +1,1 @@
+# Likelihood Ratio Based Methods for pharmacovigilance